app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "azure-cosmos>=4.9.0",
    "azure-identity>=1.23.0",
    "azure-keyvault-certificates>=4.9.0",
    "azure-keyvault-secrets>=4.9.0",
    "cryptography>=45.0.4",
    "flask==2.3",
    "gunicorn>=23.0.0",
    "requests>=2.32.4",
]
